requirements.txt
torch
torchvision
scikit-learn
pandas
numpy
matplotlib
tqdm
pillow